Change warmup scenario for execute dummy scenario (#54)
Change warmup scenario to execute dummy scenario. This way we more
accurately simulate the real behaviour of vllm inference by executing
the precise run that is happening in real inference but with dummy
config that we want to warm-up during warm-up process. No need for some
artificially create an inference scenario, as right now we are utilizing
real execution flow

---------

Signed-off-by: Agata Dobrzyniewicz <adobrzyniewicz@habana.ai>