Avoid copying dynamic slice of sampling_metadata tensors (#88)
Currently when batch has changed, we are refreshing sampling metadata on
device by copying only a slice of data.
This solution works good on GPU, but has negative impact on HPU by
causing recompilation for each unique slice.
Since we are copying here only a tensor of shape [batch_size], it is
better to avoid recompilation in cost of copying little more data.

Other solution could be to warmup each possible test case, but in this
case we would have to warmup copy of factorial of each batch_size bucket
(batch_size!), which doesn't seem like a good idea.

---------

Signed-off-by: Marcin Swiniarski <marcin.swiniarski@intel.com>