Adding dynamic swap number and defragmenter warmup (#198)
Introducing dynamic swap buckets to defragmenter, together with
defragmenter warmup.

Currently only a maximum of 32 blocks can be swapped of one iteration of
a defragmenter. This change introduces a bucketing system, which asserts
the minimal size bucket of swaps to be done in current defragmenter
iteration based on actual number of blocks, that need to be swapped.
Size of the buckets range from 8 swaps up to 512 swaps in a single
defragmenter run.

As the number of possible swap buckets grew from a single size bucket, a
warmup of defragmenter has been added. Thanks to the warmup, no
additional graph compilations connected to the defragmenter were visible
during the inference.

cherry-pick https://github.com/vllm-project/vllm-gaudi/pull/183

Signed-off-by: Krzysztof Smusz <ksmusz@habana.ai>
Co-authored-by: Marcin Swiniarski <marcin.swiniarski@intel.com>