Add HPUMultiHeadAttention with FusedSDPA (#249)
MultiheadAttention is often used in vision models. Replace naive SDPA
with FusedSDPA implementation to reduce memory usage and improve
performance in MultiHeadAttention operations.

This is part of effort for Gemma3 enablement (SW-234444)

For example, with 18 images, we get OOM (27G+ is used for processing the
vision embedding)
With this fix: Memory is used much less(less than 1G is used for the
same case)

```
image_embeds:torch.Size([18, 256, 5376])
 Before  get_multimodal_embeddings: 27.56 GiB
 After get_multimodal_embeddings: 27.09 GiB
```

Signed-off-by: Jimin Ha <jimin.ha@intel.com>
Co-authored-by: Chendi.Xue <chendi.xue@intel.com>