Fixing padded iterators in _align_and_pad (#300)
The iterators we are padding currently to match the buckets are taken
from the same object of an iterator `itertools.islice(padding_gen,
target_len)`. This is done via `itertools.repeat(padding)` being passed
to `pad_list` method, where `padding` is mentioned `islice(...)` object.

This creates an issue, where we can call `next()` on such object only
`target_len` number of times. If we are padding only one dimension e.g.
from 3 to 4, then it is sufficient for our case. But if we have to pad
more than one dimension, then all the concurrent calls to `next` on this
generator's objects do not return any value. For example, when padding
from 12 to 16 we'll only cover the 13th dimension and any `next()` calls
on 14th-16th dimensions will have already emptied iterators.

This change returns independent iterators objects for each padded
dimensions, instead of using the same iterator everywhere.

---------

Signed-off-by: Krzysztof Smusz <ksmusz@habana.ai>