Replace the deprecated logo (#564)
The old Habana logo is deprecated and should no longer be used. I
replaced it with the current logo. For the favicon, I made a slight
modification since the original logo is too long to display properly in
a browser tab, which requires a square format. As a result, the logo has
been slightly compressed, but only for this specific purpose. It looks
like this:
<img width="768" height="212" alt="image"
src="https://github.com/user-attachments/assets/68bddac8-8aef-41c0-acd1-f30d934f8de3"
/>

Signed-off-by: mhelf-intel <monika.helfer@intel.com>