Fix the spec decode the latest functionality broken (#593)
The latest vllm code will not schedule more spec decode tokens if the
latest output token is already hit the output length limit.
So there will be cases that the num draft token is 0 for some sequence
in the decoding process. This cases are not properly handled in the
current spec decode code and needs to be addressed:
1. There is a bug in our rejection_sampler for handling the case that
one or more sequence doesn't have draft tokens
2. The decode assert on spec_decode_metadata for not None but this will
happen at the end of decoding phase that all sequence reach their limit
and no draft tokens are needed.

---------

Signed-off-by: Chen Haifeng <haifeng.chen@intel.com>
Co-authored-by: Michał Kuligowski <michal.kuligowski@intel.com>