Refactor part of spec decode structure identical to vLLM (#544)
Currently a lot of spec decode related code were implemented directly in
hpu model runner and model runner will gets piled up into complexity
when more spec decode improvements and features are added.

We need to follow the same structure as vLLM implementation to put as
much spec decode related code to the corresponding proposer.

This refactor will make a basic refactor following this structure idea
but try touch as less as possible the detailed implementation logic so
that the functionality is not impacted.

---------

Signed-off-by: Chen Haifeng <haifeng.chen@intel.com>
Signed-off-by: Luca Calabria <luca.calabria@intel.com>