Spec decode: support of more than one num speculative tokens  (#609)
The main logic of support num_speculative_tokens resides in the propose
method of the refactored HpuEagleProposer class.
The generation of the first draft token is special because it can reuse
the existing attention metadata of the main model. If
num_speculative_tokens > 1, we have to do a token by token generation of
the remaining draft tokens. For each remaining draft token, we:
1. Prepare the attn metadata
2. Prepare the inputs (input token ids, positions, hidden states)
3. Call the draft model to generate a new draft token

The draft tokens generated in each iteration (including the first) stack
to get the output of the propose in the shape of [batch_size,
num_speculative_tokens].

To prepare the attn metadata, we need to pass the original block_table.
While the original block_table includes all the decodes and prefills,
when running, the decode and prefills are batched to different forwards
and we need to correctly address the block_table based on batching
information.

---------

Signed-off-by: Chen Haifeng <haifeng.chen@intel.com>