[SPEC_DECODE][BUG FIX] Fix incorrect output token with batch_size > 1 (#664)
GAUDISW-244200

Reported issue: accept rate is varying. Sometimes lower to 0.65 and
sometimes over 0.7.

Root Cause:
Generated token is mapped to wrong req (shifting to next request)

Error example:
For req4, first token 'Paris' should be output of req3
For req6, first token 'fog' should be output of req5 
```
---
Prompt: Hello, my name is
Generated text:  Xiaoyu, and I'm a student at the University of Science and Technology of China...
---
Prompt: The president of the United States is
Generated text:  nevertheless a president of the United States, and the president of the United .'...'
---
Prompt: The capital of France is
Generated text:  the city of Paris. The city of Paris is located in the north of France..'...'
---
Prompt: The future of AI is
Generated text:  Paris, 2030. The city is a hub of innovation, with AI-driven solutions tra.'...'
---
Prompt: San Francisco is know for its
Generated text:  a lot of things, but one of the most famous is the Golden Gate Bridge..'...'
---
Prompt: Facebook was created in 2004 by
Generated text:  fog, Mark Zuckerberg, and his roommates.  The name Facebook came from the n.'...'
---
Prompt: Curious George is a
Generated text:  Markov chain. The states are the different animals in the book. The tra.'...'
---
Prompt: Python 3.11 brings improvements to its
Generated text:  3rd-party libraries, including the `typing` module. One of the notable changes .'...'
```

Fix in this PR:
Main reason is that, hpu_model_runner might pad token_ids to larger size
to reduce hpu_graph possibility.
And spec decode does not trim the padding tokens in output, which result
in token generated by target_model from dummy was regarded as correct
token

After the fixing:
```
Prompt: Hello, my name is
Generated text:  Xiaoyu, and I'm a student at the University of Science and Technology of China. '...'
---
Prompt: The president of the United States is
Generated text:  the head of state and government of the United States. The president is the hea'...'
---
Prompt: The capital of France is
Generated text:  Paris. The capital of Germany is Berlin. The capital of Italy is Rome. The capital '...'
---
Prompt: The future of AI is
Generated text:  a topic that has captured the imagination of technologists, economists, an'...'
---
Prompt: San Francisco is know for its
Generated text:  fog, but the fog is not the only thing that is fog-like. The city is also known for i'...'
---
Prompt: Facebook was created in 2004 by
Generated text:  Mark Zuckerberg, and it has grown into a global social media platform w'...'
---
Prompt: Curious George is a
Generated text:  2015 American 3D computer-animated comedy film directed by Tom McCamus and'...'
---
Prompt: Python 3.11 brings improvements to its
Generated text:  standard library, including the `typing` module. One of the notable changes is'...'
``` 

Add-on:
To avoid similar issue, added accuracy test as well

---------

Signed-off-by: Chendi Xue <chendi.xue@intel.com>