[CI] GSM8K: don't override lm_eval batch_size & don't disable prefix caching by default (#640)
I'm not sure why we override the batch size - it makes the entire test
so much slower, since the requests are not pipelined (continuous
batching is effectively disabled), and I'm also not sure why we disable
prefix caching by default. vLLM keeps it on and let's do it the same
way.

---------

Signed-off-by: Konrad Zawora <kzawora@habana.ai>
Co-authored-by: Agata Dobrzyniewicz <160237065+adobrzyn@users.noreply.github.com>